Add slogan trigger command.

+++ commands/sloganme.ts
@@ -1,5 +1,6 @@
 import { Message } from "discord.js";
 import db from "../db";
+import sloganChecker from "../utilities/slogan-checker";
 
 export const command = "sloganme";
 
@@ -22,20 +23,25 @@ export const builder = {
     type: "boolean",
     default: false,
     hidden: true
+  },
+  "trigger": {
+    type: "boolean",
+    default: false,
+    hidden: true
   }
 };
 
-export async function handler({ discord, add, remove, list, count }) {
+export async function handler({ discord, add, remove, list, count, trigger }) {
   const message = discord.message as Message;
   const { sloganMembers, sloganTemplates } = db;
   if (add) {
     const newTemplate = add as string;
     if (newTemplate.length > 32) {
-      message.reply(`Slogans cannot be more than 32 characters long. Yours was ${newTemplate.length} characters long.`);
+      await message.reply(`Slogans cannot be more than 32 characters long. Yours was ${newTemplate.length} characters long.`);
       return;
     }
     if (!newTemplate.includes("{{name}}")) {
-      message.reply("Slogan template must include at least one instance of `{{name}}`.");
+      await message.reply("Slogan template must include at least one instance of `{{name}}`.");
       return;
     }
     const { resources: [existingTemplate] } = await sloganTemplates!.items.query({
@@ -43,14 +49,14 @@ export async function handler({ discord, add, remove, list, count }) {
       parameters: [{ name: "@newTemplate",  value: newTemplate }]    
     }, { partitionKey: "/_partitionKey" }).fetchAll();
     if (existingTemplate) {
-      message.reply("That slogan already exists.");
+      await message.reply("That slogan already exists.");
       return;
     }
     sloganTemplates!.items.create({
       template: newTemplate,
       _partitionKey: "/_partitionKey"
     });
-    message.reply("Slogan added!");
+    await message.reply("Slogan added!");
     return;
   }
   if (count) {
@@ -59,23 +65,28 @@ export async function handler({ discord, add, remove, list, count }) {
     }, {
       partitionKey: "/_partitionKey"
     }).fetchAll();
-    message.reply(`There are ${count} slogan templates in the database.`);
+    await message.reply(`There are ${count} slogan templates in the database.`);
     return;
   }
   if (list) {
     const { resources: templates } = await sloganTemplates!.items.readAll({ partitionKey: "/_partitionKey" }).fetchAll();
-    message.reply(templates.map(doc => `${doc.id}: ${doc.template}`).join("\n"), { split: true });
+    await message.reply(templates.map(doc => `${doc.id}: ${doc.template}`).join("\n"), { split: true });
     return;
   }
   if (remove) {
     await sloganTemplates!.item(remove, "/_partitionKey").delete();
-    message.reply("Slogan deleted.");
+    await message.reply("Slogan deleted.");
+    return;
+  }
+  if (trigger) {
+    await sloganChecker();
+    await message.reply("Slogan checker ran successfully!");
     return;
   }
   const { resource: member } = await sloganMembers!.item(message.author.id, "/_partitionKey").read();
-  if (member) {
+  if (member && member.guildId === message.guild!.id) {
     await sloganMembers!.item(message.author.id, "/_partitionKey").delete();
-    message.reply("You have been unsubscribed from sloganme!");
+    await message.reply("You have been unsubscribed from sloganme!");
     return;
   }
   await sloganMembers!.items.create({
@@ -83,5 +94,5 @@ export async function handler({ discord, add, remove, list, count }) {
     guildId: message.guild!.id,
     _partitionKey: "/_partitionKey"
   });
-  message.reply("You are now subscribed to sloganme! Every hour a D20 is rolled. If it rolls a 20 then all subscribers will see their name changed. If you'd like to contribute additional slogans, simply run `/sloganme --add \"The Power of {{name}}!\"`, making sure to include at least one instance of `{{name}}` in the slogan.");
+  await message.reply("You are now subscribed to sloganme! Every morning your nickname will change to a random slogan! If you'd like to contribute additional slogan templates, simply run `/sloganme --add \"The Power of {{name}}!\"`, making sure to include at least one instance of `{{name}}` in the slogan.");
 }
